Webinterface erweitert für Installation Squeezeboxserver (Versionsauswahl und Ajax), Update XBMC, Erweiterte Audioplayereinstellungen, Dateisystem expandieren, Statusmeldungen Sprachupdates Anpassungen auf Ubuntu 14.04

+++ max2play/application/plugins/max2play_settings/view/squeezeserver.php
@@ -22,6 +22,7 @@
 	 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */ 
 ?>										
+<script type="text/javascript" src="/js/reloadprogress.js"></script>
 
 <h1 class="entry-header">
 	<?php echo _('Squeezeserver - Status, Start & Stop') ?>
@@ -42,26 +43,45 @@
 	<form action="" method="get">
 	<input type="hidden" id="action" name="action" value="" />
 	
-	<?php if ($sp->view->pid) { ?>
-		<b><?php echo _('Status')?>:</b> <?php echo str_replace('$SERVICENAME', $sp->viewname, _('$SERVICENAME is running with processID')) ?> <b><?php echo $sp->view->pid ?></b><br /><br />
-		<a target="_blank" href="http://<?php echo preg_replace('=:.*=','', $_SERVER['HTTP_HOST']).':9000'; ?>" ><?php echo _('Start Squeezeserver Webadministration')?></a><br /><br />
-		<input type="button" value="<?php echo str_replace('$SERVICENAME', $sp->viewname, _('stop $SERVICENAME')) ?>" name="stop" onclick="document.getElementById('action').value='stop';submit();" />
-		<input type="button" value="<?php echo str_replace('$SERVICENAME', $sp->viewname, _('kill $SERVICENAME')) ?>" name="kill" onclick="document.getElementById('action').value='kill';submit();" />
-	<?php }elseif($sp->view->installed == true) { ?>
-		<b><?php echo _('Status')?>:</b> <?php echo str_replace('$SERVICENAME', $sp->viewname, _('$SERVICENAME not running')) ?><br /><br />
-		<input type="button" value="<?php echo str_replace('$SERVICENAME', $sp->viewname, _('start $SERVICENAME')) ?>" name="start" onclick="document.getElementById('action').value='start';submit();" />
-		
-	<?php }else { ?>				
-			<b><?php echo str_replace('$NAME', $sp->viewname, _('$NAME start installation')) ?>:</b> 
-			<input type="button" value="<?php echo str_replace('$NAME', $sp->viewname, _('$NAME start installation')) ?>" name="install" onclick="document.getElementById('action').value='install';submit();" />
-			<br /><?php echo _('The installation takes about 5 to 10 minutes depending on your internet connection. At first it downloads the package from http://downloads.slimdevices.com/ and afterwards it installs the package. You may reload this page by clicking the button again to see the status of the install process.')?>	
-	<?php } ?>
-	
+	<?php if($sp->view->installed == true) { ?>
+		<?php if ($sp->view->pid) { ?>
+			<b><?php echo _('Status')?>:</b> <?php echo str_replace('$SERVICENAME', $sp->viewname, _('$SERVICENAME is running with processID')) ?> <b><?php echo $sp->view->pid ?></b><br /><br />
+			<a target="_blank" href="http://<?php echo preg_replace('=:.*=','', $_SERVER['HTTP_HOST']).':9000'; ?>" ><?php echo _('Start Squeezeserver Webadministration')?></a><br /><br />
+			<input type="button" value="<?php echo str_replace('$SERVICENAME', $sp->viewname, _('stop $SERVICENAME')) ?>" name="stop" onclick="document.getElementById('action').value='stop';submit();" />
+			<input type="button" value="<?php echo str_replace('$SERVICENAME', $sp->viewname, _('kill $SERVICENAME')) ?>" name="kill" onclick="document.getElementById('action').value='kill';submit();" />
+		<?php }else { ?>		
+			<b><?php echo _('Status')?>:</b> <?php echo str_replace('$SERVICENAME', $sp->viewname, _('$SERVICENAME not running')) ?><br /><br />
+			<input type="button" value="<?php echo str_replace('$SERVICENAME', $sp->viewname, _('start $SERVICENAME')) ?>" name="start" onclick="document.getElementById('action').value='start';submit();" />			
+		<?php } ?>
 		<br /><br />
 		<b><?php echo str_replace('$SERVICENAME', $sp->viewname, _('Autostart $SERVICENAME')) ?>:</b><br />
 		<input type="checkbox" <?php if($sp->view->autostart) echo "checked"; ?> value="1" name="autostart" />
 		<input type="button" value="<?php echo _('save') ?>" name="save" onclick="document.getElementById('action').value='save';submit();" />
+		<br /><br />
+	<?php } ?>				
+							
+	<p class="ui-state-default ui-corner-all" style="padding:4px;margin-bottom:1em;">
+		<span class="ui-icon ui-icon-gear" style="float:left; margin:-2px 5px 0 0;"></span>
+		<?php echo str_replace('$NAME', $sp->viewname, _('$NAME start installation')) ?></p>
+	<select name="lmsversion" onclick="copytodownloadurl(this);">
+	<?php foreach($sp->lmsversions as $key => $value) { ?>
+		<option value="<?php echo $value ?>"><?php echo $key ?></option>
+	<?php } ?>
+	</select>
+	<script type="text/javascript">
+		function copytodownloadurl(item){
+			document.getElementById("downloadurl").value = item.options[item.selectedIndex].value;
+			document.getElementById("downloadurl").style = "width:500px";
+		}
+	</script>
+	<input type="button" value="<?php echo _('Show available Versions') ?>" name="fetchLMS" onclick="document.getElementById('action').value='showavailablelms';submit();" /><br /><br />
+	 
+	<?php echo _('Alternative other source from Slimdevices (package must be .deb)') ?>: 
+	<input id="downloadurl" type="text" value="" name="downloadurl" style="width:200px;" /><br /><br /> 
 	
+	<input type="button" value="<?php echo str_replace('$NAME', $sp->viewname, _('$NAME start installation')) ?>" name="install" onclick="document.getElementById('action').value='install';submit();" />
+	<br /><br /> <?php echo _('The installation takes about 5 to 10 minutes depending on your internet connection. At first it downloads the package from http://downloads.slimdevices.com/ and afterwards it installs the package. You may reload this page by clicking the button again to see the status of the install process.')?>			
+		
 	</form>
 	
 	<br /><br />
